Added steamdeck-retroarch-download-all-cores.sh
This is a script to get _all_ available libretro cores installed into
the Steam version of RetroArch.

Put it in a cronjob that runs frequently; it'll notice if the version
hasn't changed (that is, Steam hasn't updated RetroArch since last time)
and will quit immediately, only downloading/updating cores for new Steam
releases.

This was meant for use on the Steam Deck, but it works on x86-64 desktop
Linux, too. Sorry, Windows users.