steamdeck-retroarch-download-all-cores.sh: Install some needed BIOSes.
This currently pulls in the PlayStation 2 BIOS files (for pcsx2) and
the MSX BIOS files (for fMSX, in case you want the original version
of Metal Gear, heh), and a few others that may or may not be required.

These pull from the Internet Archive. Others can be added as necessary.