Added '$': A script to let you paste in command lines from webpages.
Every webpage that has a HOWTO about some Linux thing always lists command
lines you have to punch in like this:

```
$ command1 --some args
$ command2 other ./thing
$ sudo rm -rf /
```

And you can't just paste it from the page because they started each line with
a dollar sign to make sure you know this is a command line thing, or whatever.

So hey, let's make a script called '$' and put it in our $PATH.

The only thing it does is exec its command line arguments, but it'll let you
mostly get away with pasting multiple lines into your terminal at once, or
at least not having to do gymnastics to highlight everything but the '$' on
the page.

Obviously this won't work for _everything_ you can do at a command line, but
mostly you're just running a bunch of other programs in these cases, and this
helps with that.

(disclaimer: don't blindly run commands you see on webpages. Think for a
second before pasting it into a terminal, whether you use this script or not!)